# milktea/skills/milktea-openapi.yaml
openapi: 3.1.0
info:
  title: MilkTea HTTP API — thiết kế một quán
  version: 2.0.0
  description: 'Hợp đồng thiết kế để triển khai; không xác nhận endpoint đã có code. Domain: domain_system.md; HTTP
    envelope: system_architecture.md; security: author_rule.md. Không mô tả STOMP trong OpenAPI. QR read-only; session
    mở khi order/CASHIER open; thanh toán linh hoạt; chỉ CASHIER close. x-allowed-account-roles chỉ ràng buộc nhánh
    xác thực tài khoản, KHÔNG phải điều kiện role cho guest. Guest được cho phép theo nhánh qrAccess/tableAccess/counterAccess
    và phạm vi resource tương ứng; luôn kiểm Domain. Tooling OpenAPI không tự enforce role hay ownership.

    Nguồn: báo cáo MilkTea cập nhật 05/10/2026, 41 UC/T01–T60. Hợp đồng thiết kế; chưa xác nhận các endpoint đã
    được triển khai. v2 bổ sung UC23 thông tin quán, UC30 ngưỡng theo vị trí, UC39 sự cố/tiêu hao pha bù; version
    BIGINT truyền chuỗi; giữ giỏ/context trong kết quả đặt bàn.'
servers:
- url: http://localhost:8080
  description: Ví dụ local; dùng SERVER_PORT và origin thực của project.
paths:
...
  /api/v1/customer/orders:
    get:
      operationId: listOwnOrders
      summary: Lịch sử đơn đúng account
      description: Lịch sử đơn đúng account
      tags:
      - Customer
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderView'
                  meta:
                    $ref: '#/components/schemas/PageMeta'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
        description: ''
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        description: ''
      - name: sort
        in: query
        required: false
        schema:
          type: string
          maxLength: 255
        description: Allowlist sort field; default createdAt,desc.
      x-allowed-account-roles:
      - CUSTOMER
      x-use-cases:
      - UC20
...
  /api/v1/orders/table:
    post:
      operationId: createTableOrder
      summary: Đặt món tại bàn; mở phiên chỉ khi tạo đơn đầu tiên
      description: QR/table credential xác minh đúng tableId. expectedSessionId=null chỉ khi trang chưa có phiên;
        nếu hiện trạng khác trả 409. Khi mở mới dùng remainingCartItems (mặc định rỗng) lưu món nháp chưa gửi; khi
        phiên đã mở không cho remainingCartItems, cartVersion bắt buộc và items lấy từ giỏ version đó, không vượt
        lượng giỏ. Tạo đơn+hóa đơn+snapshot, giảm giỏ đúng lượng và tăng version cùng commit. Không có tiền/role/accountId
        do client tự gán. Có JWT cookie hợp lệ thì ghi account owner, nhưng vẫn cần quyền bàn. replacementOfOrderId
        chỉ khi được phép xem đơn gốc đã hủy trong đúng phạm vi. Không yêu cầu trả trước; không nhận tiền bằng API
        này. Response session trả giỏ và version sau commit để frontend dùng bản chốt. cartVersion dùng chuỗi BIGINT
        hoặc null.
      tags:
      - Orders
      security:
      - qrAccess: []
      - tableAccess: []
      responses:
        '201':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/OrderCreatedView'
                  meta:
                    type: 'null'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: X-CSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
          maxLength: 1024
        description: Required for browser mutations. Fetch /api/v1/auth/csrf; never read account JWT.
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 64
        description: Unpredictable client request key; same scope+payload replay returns same resource, changed
          payload returns 409.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableOrderRequest'
      x-use-cases:
      - UC08
      - UC17
  /api/v1/orders/counter:
    post:
      operationId: createGuestCounterOrder
      summary: Đặt đơn quầy cho guest hoặc CUSTOMER
      description: Công khai có CSRF; JWT hợp lệ gắn owner account, guest trả counterOrderToken RAM only. Không
        nhận role/giá từ client. Không tự gán bàn. Khi có replacementOfOrderId, bắt buộc credential owner account
        hoặc X-Counter-Token đúng đơn gốc đã hủy; không chỉ tin mã đơn.
      tags:
      - Orders
      security: []
      responses:
        '201':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/OrderCreatedView'
                  meta:
                    type: 'null'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: X-Counter-Token
        in: header
        required: false
        schema:
          type: string
          maxLength: 512
        description: Bắt buộc khi guest tạo đơn thay thế từ đơn quầy cũ; không cần khi tạo mới thông thường.
      - name: X-CSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
          maxLength: 1024
        description: Required for browser mutations. Fetch /api/v1/auth/csrf; never read account JWT.
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 64
        description: Unpredictable client request key; same scope+payload replay returns same resource, changed
          payload returns 409.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CounterOrderRequest'
      x-use-cases:
      - UC09
      - UC17
  /api/v1/orders/{orderId}:
    get:
      operationId: readOrder
      summary: Xem đơn đúng owner/phiên/credential hoặc staff
      description: Xem đơn đúng owner/phiên/credential hoặc staff
      tags:
      - Orders
      security:
      - accountCookie: []
      - bearerAuth: []
      - tableAccess: []
      - counterAccess: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/OrderView'
                  meta:
                    type: 'null'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        description: ''
      x-allowed-account-roles:
      - CUSTOMER
      - CASHIER
      - KITCHEN
      - ADMIN
      x-use-cases:
      - UC10
...
  /api/v1/cashier/orders:
    get:
      operationId: listCashierOrders
      summary: Quầy xem danh sách đơn
      description: Quầy xem danh sách đơn
      tags:
      - Cashier
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderView'
                  meta:
                    $ref: '#/components/schemas/PageMeta'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
        description: ''
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        description: ''
      - name: sort
        in: query
        required: false
        schema:
          type: string
          maxLength: 255
        description: Allowlist sort field; default createdAt,desc.
      - name: status
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/OrderStatus'
        description: ''
      x-allowed-account-roles:
      - CASHIER
      x-use-cases:
      - UC10
    post:
      operationId: createCashierOrder
      summary: Thu ngân tạo đơn quầy và có thể chọn bàn
      description: Source COUNTER. tableId null giữ quầy; có tableId gán/mở phiên qua thao tác CASHIER, expectedSessionId
        phải khớp. Không tự ghi thu.
      tags:
      - Cashier
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '201':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/OrderCreatedView'
                  meta:
                    type: 'null'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: X-CSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
          maxLength: 1024
        description: Required for browser mutations. Fetch /api/v1/auth/csrf; never read account JWT.
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 64
        description: Unpredictable client request key; same scope+payload replay returns same resource, changed
          payload returns 409.
      x-allowed-account-roles:
      - CASHIER
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashierOrderRequest'
      x-use-cases:
      - UC09
      - UC17
  /api/v1/cashier/orders/{orderId}/confirm:
    post:
      operationId: confirmOrder
      summary: Thu ngân xác nhận xuống bếp
      description: CHO_XAC_NHAN→CHO_CHE_BIEN, không kiểm điều kiện đã thu. Không cho CASHIER complete.
      tags:
      - Cashier
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/OrderView'
                  meta:
                    type: 'null'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        description: ''
      - name: X-CSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
          maxLength: 1024
        description: Required for browser mutations. Fetch /api/v1/auth/csrf; never read account JWT.
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 64
        description: Unpredictable client request key; same scope+payload replay returns same resource, changed
          payload returns 409.
      x-allowed-account-roles:
      - CASHIER
      x-use-cases:
      - UC11
  /api/v1/cashier/orders/{orderId}/assign-table:
    post:
      operationId: assignCounterOrderToTable
      summary: Gán đơn quầy vào phiên bàn
      description: Giữ đơn/hóa đơn/thu/status/source COUNTER. Mở phiên khi chưa có, dùng phiên đúng khi đã mở. Không
        di chuyển khỏi phiên khác đang hoạt động.
      tags:
      - Cashier
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/OrderView'
                  meta:
                    type: 'null'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        description: ''
      - name: X-CSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
          maxLength: 1024
        description: Required for browser mutations. Fetch /api/v1/auth/csrf; never read account JWT.
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 64
        description: Unpredictable client request key; same scope+payload replay returns same resource, changed
          payload returns 409.
      x-allowed-account-roles:
      - CASHIER
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignTableRequest'
      x-use-cases:
      - UC40
...
  /api/v1/kitchen/orders:
    get:
      operationId: listKitchenOrders
      summary: Bếp xem đơn đã xác nhận kể cả chưa thu
      description: Bếp xem đơn đã xác nhận kể cả chưa thu
      tags:
      - Kitchen
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderView'
                  meta:
                    $ref: '#/components/schemas/PageMeta'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
        description: ''
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        description: ''
      - name: sort
        in: query
        required: false
        schema:
          type: string
          maxLength: 255
        description: Allowlist sort field; default createdAt,desc.
      - name: status
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/OrderStatus'
        description: ''
      x-allowed-account-roles:
      - KITCHEN
      x-use-cases:
      - UC10
      - UC13
  /api/v1/kitchen/orders/{orderId}/start:
    post:
      operationId: startOrderPreparation
      summary: Bếp bắt đầu pha toàn đơn
      description: CHO_CHE_BIEN→DANG_CHE_BIEN; không pending cancellation; không payment prerequisite.
      tags:
      - Kitchen
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/OrderView'
                  meta:
                    type: 'null'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        description: ''
      - name: X-CSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
          maxLength: 1024
        description: Required for browser mutations. Fetch /api/v1/auth/csrf; never read account JWT.
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 64
        description: Unpredictable client request key; same scope+payload replay returns same resource, changed
          payload returns 409.
      x-allowed-account-roles:
      - KITCHEN
      x-use-cases:
      - UC13
  /api/v1/kitchen/orders/{orderId}/complete:
    post:
      operationId: completeOrderPreparation
      summary: Chỉ bếp hoàn thành và ghi tiêu hao
      description: DANG_CHE_BIEN→HOAN_THANH, inventory cùng transaction REQUIRED; thiếu tồn rollback, không trừ
        lặp, không close bàn và không yêu cầu đã thu.
      tags:
      - Kitchen
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/OrderView'
                  meta:
                    type: 'null'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        description: ''
      - name: X-CSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
          maxLength: 1024
        description: Required for browser mutations. Fetch /api/v1/auth/csrf; never read account JWT.
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 64
        description: Unpredictable client request key; same scope+payload replay returns same resource, changed
          payload returns 409.
      x-allowed-account-roles:
      - KITCHEN
      x-use-cases:
      - UC14
...
  /api/v1/admin/reports/orders:
    get:
      operationId: readOrderReport
      summary: Báo cáo số đơn theo trạng thái
      description: Báo cáo số đơn theo trạng thái
      tags:
      - Admin
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/OrderReportView'
                  meta:
                    type: 'null'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: fromDate
        in: query
        required: true
        schema:
          type: string
          format: date
        description: ''
      - name: toDate
        in: query
        required: true
        schema:
          type: string
          format: date
        description: ''
      x-allowed-account-roles:
      - ADMIN
      x-use-cases:
      - UC36
...
components:
  securitySchemes:
    accountCookie:
      type: apiKey
      in: cookie
      name: milktea_access_token
      description: JWT account HttpOnly; role/state/version enforced server-side.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tableAccess:
      type: apiKey
      in: header
      name: X-Table-Token
      description: Opaque access scoped to one currently OPEN TableSession, not an account role.
    counterAccess:
      type: apiKey
      in: header
      name: X-Counter-Token
      description: Opaque access scoped to one counter order; client keeps only in current page RAM.
    qrAccess:
      type: apiKey
      in: header
      name: X-QR-Code
      description: Opaque QR code identifies a table; never a staff credential.
  schemas:
    Id:
      type: string
      pattern: ^[1-9][0-9]*$
      example: '5'
...
    OrderStatus:
      type: string
      enum:
      - CHO_XAC_NHAN
      - CHO_CHE_BIEN
      - DANG_CHE_BIEN
      - HOAN_THANH
      - DA_HUY
    OrderSource:
      type: string
      enum:
      - TABLE
      - COUNTER
...
    PageMeta:
      type: object
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 0
        totalElements:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
      additionalProperties: false
      required:
      - page
      - size
      - totalElements
      - totalPages
...
    OrderView:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        source:
          $ref: '#/components/schemas/OrderSource'
        sessionId:
          anyOf:
          - $ref: '#/components/schemas/Id'
          - type: 'null'
        status:
          $ref: '#/components/schemas/OrderStatus'
        createdAt:
          type: string
          format: date-time
        completedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemView'
        invoice:
          $ref: '#/components/schemas/InvoiceView'
        payments:
          type: array
          items:
            $ref: '#/components/schemas/PaymentView'
        cancellations:
          type: array
          items:
            $ref: '#/components/schemas/CancellationView'
        refunds:
          type: array
          items:
            $ref: '#/components/schemas/RefundView'
        replacementOfOrderId:
          anyOf:
          - $ref: '#/components/schemas/Id'
          - type: 'null'
      additionalProperties: false
      required:
      - id
      - source
      - sessionId
      - status
      - createdAt
      - completedAt
      - items
      - invoice
      - payments
      - cancellations
      - refunds
      - replacementOfOrderId
...
    CounterOrderRequest:
      type: object
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/CartItemInput'
        replacementOfOrderId:
          anyOf:
          - $ref: '#/components/schemas/Id'
          - type: 'null'
      additionalProperties: false
      required:
      - items
...
  responses:
    Error400:
      description: HTTP 400; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error401:
      description: HTTP 401; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error403:
      description: HTTP 403; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error404:
      description: HTTP 404; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error409:
      description: HTTP 409; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Error500:
      description: HTTP 500; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
- accountCookie: []
- bearerAuth: []
...